# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for analyzing STL files.",
        "title": "STL File Analysis API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/stl/triangles": {
            "post": {
                "description": "Receives an STL file and returns the number of triangles and the surface area",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stl"
                ],
                "summary": "Get number of triangles and surface area of an STL file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "STL file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.HTTPResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.TrianglesHTTPResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.HTTPResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.TrianglesHTTPResponse": {
            "type": "object",
            "properties": {
                "number_of_triangles": {
                    "type": "integer"
                },
                "surface_area": {
                    "type": "number"
                }
            }
        },
        "utils.HTTPResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}